Cleaned up messages. Prevent multiple CTRL-C

+++ metaflow/plugins/aws/batch/batch.py
@@ -29,16 +29,7 @@ class Batch(object):
         self.metadata = metadata
         self.environment = environment
         self._client = BatchClient()
-        def f():
-            print("Atexit executing")
-            if hasattr(self, 'job'):
-                print("Killing")
-                self.job.kill()
-                print("Done killing")
-            else:
-                print("No job")
-        atexit.register(f)
-        #atexit.register(lambda : self.job.kill() if hasattr(self, 'job') else None)
+        atexit.register(lambda: self.job.kill() if hasattr(self, 'job') else None)
 
     def _command(self, code_package_url, environment, step_name, step_cli):
         cmds = environment.get_package_commands(code_package_url)
